+++ emulator/cpu/isa/int.h
@@ -1,6 +1,6 @@
 #ifndef __INT__
 #define __INT__
 
-	make_helper(INT);
+	make_helper(int_);
 
 #endif
